requirements.txt
Django
pillow
djangorestframework
django-reset-migrations
django-mptt
django-cors-headers
geopy
psycopg2-binary
django-ipware
requests